+++ src/pages/project-editor/AdminEditSection/MembersInput.tsx
@@ -6,90 +6,96 @@ import { IoPersonOutline } from 'react-icons/io5';
 interface MembersInputProps {
   teamMembers: TeamMember[];
   onMemberAdd: (newMemberName: string) => void;
-  onMemberRemove: (index: number) => void;
+  onMemberRemove: (teamMemberId: number) => void;
 }
 
 const MAX_MEMBERS = 6;
 
 const MembersInput = ({ teamMembers, onMemberAdd, onMemberRemove }: MembersInputProps) => {
-  const [inputs, setInputs] = useState<string[]>(
-    Array.from({ length: MAX_MEMBERS }, (_, i) => teamMembers[i]?.teamMemberName || ''),
-  );
-  const [added, setAdded] = useState<boolean[]>(
-    Array.from({ length: MAX_MEMBERS }, (_, i) => !!teamMembers[i]?.teamMemberName),
-  );
-
-  const handleInputChange = (idx: number, value: string) => {
-    setInputs((inputs) => inputs.map((v, i) => (i === idx ? value : v)));
-  };
-
-  const handleAdd = (idx: number) => {
-    if (!inputs[idx].trim()) return;
-    setAdded((added) => added.map((a, i) => (i === idx ? true : a)));
-    onMemberAdd(inputs[idx].trim());
-  };
-
-  const handleRemove = (idx: number) => {
-    setInputs((inputs) => inputs.map((v, i) => (i === idx ? '' : v)));
-    setAdded((added) => added.map((a, i) => (i === idx ? false : a)));
-    onMemberRemove(idx);
-  };
+  const [inputs, setInputs] = useState<Record<string, string>>({});
+  const emptySlotsCount = MAX_MEMBERS - teamMembers.length;
 
   useEffect(() => {
-    setInputs(Array.from({ length: MAX_MEMBERS }, (_, i) => teamMembers[i]?.teamMemberName || ''));
-    setAdded(Array.from({ length: MAX_MEMBERS }, (_, i) => !!teamMembers[i]?.teamMemberName));
+    const newInputs: Record<string, string> = {};
+    teamMembers.forEach((member) => {
+      newInputs[member.teamMemberId.toString()] = member.teamMemberName;
+    });
+    setInputs(newInputs);
   }, [teamMembers]);
 
+  const [newInputs, setNewInputs] = useState<string[]>(Array(emptySlotsCount).fill(''));
+
+  const handleNewInputChange = (idx: number, value: string) => {
+    setNewInputs((prev) => {
+      const copy = [...prev];
+      copy[idx] = value;
+      return copy;
+    });
+  };
+
+  const handleAddNewMember = (idx: number) => {
+    const trimmed = newInputs[idx].trim();
+    if (!trimmed) return;
+    onMemberAdd(trimmed);
+    setNewInputs((prev) => {
+      const copy = [...prev];
+      copy[idx] = '';
+      return copy;
+    });
+  };
+
   return (
-    <div className="flex flex-col gap-5 text-sm sm:flex-row sm:items-start sm:gap-10">
-      <div className="text-midGray flex w-25 gap-1">
-        <span className="mr-1 text-red-500">*</span>
-        <span className="w-full">팀원</span>
-      </div>
-      <div className="flex flex-1 flex-col">
-        <div className="grid w-full grid-cols-1 gap-3 text-sm lg:grid-cols-2">
-          {Array.from({ length: MAX_MEMBERS }).map((_, idx) => (
-            <div
-              key={idx}
-              className={`relative flex items-center gap-1 rounded border px-4 py-2 text-sm ${added[idx] ? 'border-mainGreen' : 'border-gray-300'}`}
-            >
-              <IoPersonOutline className={`${added[idx] ? 'text-mainGreen' : 'text-lightGray'} mr-2`} size={20} />
-              <input
-                type="text"
-                className="placeholder:text-lightGray flex-1 bg-transparent focus:outline-none"
-                placeholder="팀원명을 입력해주세요."
-                value={inputs[idx]}
-                onChange={(e) => handleInputChange(idx, e.target.value)}
-                disabled={added[idx]}
-                onKeyDown={(e) => {
-                  if (e.key === 'Enter' && !added[idx] && inputs[idx].trim()) handleAdd(idx);
-                }}
-              />
-              {!added[idx] && (
-                <button
-                  type="button"
-                  onClick={() => handleAdd(idx)}
-                  className={`focus:outline-none ${inputs[idx].trim() ? 'text-mainGreen' : 'text-lightGray'}`}
-                  disabled={!inputs[idx].trim()}
-                  aria-label="Add member"
-                >
-                  <IoIosAdd size={24} />
-                </button>
-              )}
-              {added[idx] && (
-                <button
-                  type="button"
-                  onClick={() => handleRemove(idx)}
-                  className="text-mainGreen focus:outline-none"
-                  aria-label="Remove member"
-                >
-                  <IoIosRemove size={24} />
-                </button>
-              )}
-            </div>
-          ))}
+    <div className="grid w-full grid-cols-1 gap-3 text-sm lg:grid-cols-2">
+      {teamMembers.map((member) => (
+        <div
+          key={member.teamMemberId}
+          className="border-mainGreen relative flex items-center gap-1 rounded border px-4 py-2 text-sm"
+        >
+          <IoPersonOutline className="text-mainGreen mr-2" size={20} />
+          <input
+            type="text"
+            className="placeholder:text-lightGray flex-1 bg-transparent focus:outline-none"
+            value={inputs[member.teamMemberId.toString()] || ''}
+            disabled // 기존 멤버는 수정 불가 (원하면 삭제 가능)
+            readOnly
+          />
+          <button
+            type="button"
+            onClick={() => onMemberRemove(member.teamMemberId)}
+            className="text-mainGreen"
+            aria-label={`Remove member ${member.teamMemberName}`}
+          >
+            <IoIosRemove size={24} />
+          </button>
         </div>
-      </div>
+      ))}
+      {Array.from({ length: emptySlotsCount }).map((_, idx) => (
+        <div
+          key={`new-${idx}`}
+          className="relative flex items-center gap-1 rounded border border-gray-300 px-4 py-2 text-sm"
+        >
+          <IoPersonOutline className="text-lightGray mr-2" size={20} />
+          <input
+            type="text"
+            placeholder="팀원명을 입력해주세요."
+            className="placeholder:text-lightGray flex-1 bg-transparent focus:outline-none"
+            value={newInputs[idx]}
+            onChange={(e) => handleNewInputChange(idx, e.target.value)}
+            onKeyDown={(e) => {
+              if (e.key === 'Enter' && newInputs[idx].trim()) handleAddNewMember(idx);
+            }}
+          />
+          <button
+            type="button"
+            onClick={() => handleAddNewMember(idx)}
+            className={`${newInputs[idx].trim() ? 'text-mainGreen' : 'text-lightGray'}`}
+            disabled={!newInputs[idx].trim()}
+            aria-label="Add member"
+          >
+            <IoIosAdd size={24} />
+          </button>
+        </div>
+      ))}
     </div>
   );
 };
